Restructure app into client side and server side folders
